Display the PDF instead of a PNG in the HTML page.
This is because the `convert` command to generate the PNG doesn't work
in the Docker container case.

Also, I initially thought it was nicer to show a PNG, but having the
ability to scroll the whole PDF is not bad. And in my little example,
the PDF is actually a smaller file than the PNG.